+++ httpx_auth/authentication.py
@@ -169,6 +169,7 @@ class OAuth2ResourceOwnerPasswordCredentials(httpx.Auth, SupportMultiAuth):
             raise Exception("header_value parameter must contains {token}.")
 
         self.token_field_name = kwargs.pop("token_field_name", None) or "access_token"
+        self.early_expiry = 30.0
 
         # Time is expressed in seconds
         self.timeout = int(kwargs.pop("timeout", None) or 60)
@@ -194,7 +195,9 @@ class OAuth2ResourceOwnerPasswordCredentials(httpx.Auth, SupportMultiAuth):
         self, request: httpx.Request
     ) -> Generator[httpx.Request, httpx.Response, None]:
         token = OAuth2.token_cache.get_token(
-            self.state, on_missing_token=self.request_new_token
+            self.state,
+            early_expiry=self.early_expiry,
+            on_missing_token=self.request_new_token,
         )
         request.headers[self.header_name] = self.header_value.format(token=token)
         yield request
@@ -250,6 +253,7 @@ class OAuth2ClientCredentials(httpx.Auth, SupportMultiAuth):
             raise Exception("header_value parameter must contains {token}.")
 
         self.token_field_name = kwargs.pop("token_field_name", None) or "access_token"
+        self.early_expiry = 30.0
 
         # Time is expressed in seconds
         self.timeout = int(kwargs.pop("timeout", None) or 60)
@@ -272,7 +276,9 @@ class OAuth2ClientCredentials(httpx.Auth, SupportMultiAuth):
         self, request: httpx.Request
     ) -> Generator[httpx.Request, httpx.Response, None]:
         token = OAuth2.token_cache.get_token(
-            self.state, on_missing_token=self.request_new_token
+            self.state,
+            early_expiry=self.early_expiry,
+            on_missing_token=self.request_new_token,
         )
         request.headers[self.header_name] = self.header_value.format(token=token)
         yield request
@@ -350,6 +356,7 @@ class OAuth2AuthorizationCode(httpx.Auth, SupportMultiAuth, BrowserAuth):
             raise Exception("header_value parameter must contains {token}.")
 
         self.token_field_name = kwargs.pop("token_field_name", None) or "access_token"
+        self.early_expiry = 30.0
 
         username = kwargs.pop("username", None)
         password = kwargs.pop("password", None)
@@ -405,7 +412,9 @@ class OAuth2AuthorizationCode(httpx.Auth, SupportMultiAuth, BrowserAuth):
         self, request: httpx.Request
     ) -> Generator[httpx.Request, httpx.Response, None]:
         token = OAuth2.token_cache.get_token(
-            self.state, on_missing_token=self.request_new_token
+            self.state,
+            early_expiry=self.early_expiry,
+            on_missing_token=self.request_new_token,
         )
         request.headers[self.header_name] = self.header_value.format(token=token)
         yield request
@@ -491,6 +500,7 @@ class OAuth2AuthorizationCodePKCE(httpx.Auth, SupportMultiAuth, BrowserAuth):
             raise Exception("header_value parameter must contains {token}.")
 
         self.token_field_name = kwargs.pop("token_field_name", None) or "access_token"
+        self.early_expiry = 30.0
 
         # As described in https://tools.ietf.org/html/rfc6749#section-4.1.2
         code_field_name = kwargs.pop("code_field_name", "code")
@@ -553,7 +563,9 @@ class OAuth2AuthorizationCodePKCE(httpx.Auth, SupportMultiAuth, BrowserAuth):
         self, request: httpx.Request
     ) -> Generator[httpx.Request, httpx.Response, None]:
         token = OAuth2.token_cache.get_token(
-            self.state, on_missing_token=self.request_new_token
+            self.state,
+            early_expiry=self.early_expiry,
+            on_missing_token=self.request_new_token,
         )
         request.headers[self.header_name] = self.header_value.format(token=token)
         yield request
@@ -675,6 +687,8 @@ class OAuth2Implicit(httpx.Auth, SupportMultiAuth, BrowserAuth):
                 "id_token" if "id_token" == response_type else "access_token"
             )
 
+        self.early_expiry = 30.0
+
         authorization_url_without_nonce = _add_parameters(
             self.authorization_url, kwargs
         )
@@ -702,6 +716,7 @@ class OAuth2Implicit(httpx.Auth, SupportMultiAuth, BrowserAuth):
     ) -> Generator[httpx.Request, httpx.Response, None]:
         token = OAuth2.token_cache.get_token(
             self.state,
+            early_expiry=self.early_expiry,
             on_missing_token=oauth2_authentication_responses_server.request_new_grant,
             grant_details=self.grant_details,
         )